Server: Rework main loop to use libevent and add scheduler subsystem.
The task/event scheduler is supposed to be a simple way to execute
handle multiple long running tasks in parallel on a predefined nuber
of threads. The code is not used right now, but will be in audio
converter module.

The task scheduler also provides a simple way to register an immediate
callback function executed directly from the main event loop.

Signed-off-by: Peter Tworek <tworaz666@gmail.com>